+++ servicetalk-loadbalancer-experimental/src/main/java/io/servicetalk/loadbalancer/P2CSelector.java
@@ -19,6 +19,8 @@ import io.servicetalk.client.api.LoadBalancedConnection;
 import io.servicetalk.client.api.LoadBalancer;
 import io.servicetalk.concurrent.api.Single;
 import io.servicetalk.context.api.ContextMap;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Random;
@@ -36,24 +38,36 @@ import javax.annotation.Nullable;
 final class P2CSelector<ResolvedAddress, C extends LoadBalancedConnection>
         extends BaseHostSelector<ResolvedAddress, C> {
 
-    private final List<? extends Host<ResolvedAddress, C>> hosts;
+    private static final Logger LOGGER = LoggerFactory.getLogger(P2CSelector.class);
+
+    private static double ACCEPTABLE_PERCENT_ERROR = 0.01;
+    private static final EntrySelector EMPTY_SELECTOR = new EqualWeightTable(0);
+
     @Nullable
     private final Random random;
+    private final boolean supportWeights;
+    private final List<? extends Host<ResolvedAddress, C>> hosts;
+    private final EntrySelector entrySelector;
     private final int maxEffort;
     private final boolean failOpen;
+    private final String targetResource;
 
-    P2CSelector(List<? extends Host<ResolvedAddress, C>> hosts, final String targetResource, final int maxEffort,
-                final boolean failOpen, @Nullable final Random random) {
+    P2CSelector(List<? extends Host<ResolvedAddress, C>> hosts, final String targetResource,
+                       final boolean supportWeights, final int maxEffort, final boolean failOpen,
+                       @Nullable final Random random) {
         super(hosts, targetResource);
+        this.supportWeights = supportWeights;
         this.hosts = hosts;
+        this.entrySelector = supportWeights ? buildAliasTable(hosts) : new EqualWeightTable(hosts.size());
         this.maxEffort = maxEffort;
         this.failOpen = failOpen;
+        this.targetResource = targetResource;
         this.random = random;
     }
 
     @Override
     public HostSelector<ResolvedAddress, C> rebuildWithHosts(List<? extends Host<ResolvedAddress, C>> hosts) {
-        return new P2CSelector<>(hosts, getTargetResource(), maxEffort, failOpen, random);
+        return new P2CSelector<>(hosts, targetResource, supportWeights, maxEffort, failOpen, random);
     }
 
     @Override
@@ -79,25 +93,20 @@ final class P2CSelector<ResolvedAddress, C extends LoadBalancedConnection>
                 }
                 return noActiveHostsFailure(hosts);
             default:
-                return p2c(size, hosts, getRandom(), selector, forceNewConnectionAndReserve, context);
+                return p2c(hosts, getRandom(), selector, forceNewConnectionAndReserve, context);
         }
     }
 
-    private Single<C> p2c(int size, List<? extends Host<ResolvedAddress, C>> hosts, Random random,
+    private Single<C> p2c(List<? extends Host<ResolvedAddress, C>> hosts, Random random,
                           Predicate<C> selector, boolean forceNewConnectionAndReserve,
                           @Nullable ContextMap contextMap) {
         // If there are only two hosts we only try once since there is no chance we'll select different hosts
         // on further iterations.
         Host<ResolvedAddress, C> failOpenHost = null;
         for (int j = hosts.size() == 2 ? 1 : maxEffort; j > 0; j--) {
-            // Pick two random indexes that don't collide. Limit the range on the second index to 1 less than
-            // the max value so that if there is a collision we can safety increment. We also increment if
-            // i2 > i1 to avoid bias toward lower numbers since we limited the range by 1.
-            final int i1 = random.nextInt(size);
-            int i2 = random.nextInt(size - 1);
-            if (i2 >= i1) {
-                ++i2;
-            }
+            // Pick two random indexes that don't collide.
+            final int i1 = entrySelector.firstEntry(random);
+            final int i2 = entrySelector.secondEntry(random, i1);
 
             Host<ResolvedAddress, C> t1 = hosts.get(i1);
             Host<ResolvedAddress, C> t2 = hosts.get(i2);
@@ -157,5 +166,173 @@ final class P2CSelector<ResolvedAddress, C extends LoadBalancedConnection>
 
     private Random getRandom() {
         return random == null ? ThreadLocalRandom.current() : random;
+    }
+
+    private EntrySelector buildAliasTable(List<? extends Host<?, ?>> hosts) {
+        if (hosts.isEmpty()) {
+            return EMPTY_SELECTOR;
+        }
+
+        double[] probs = new double[hosts.size()];
+        boolean allSameProbability = true;
+        double pTotal = 0;
+        for (int i = 0; i < hosts.size(); i++) {
+            final double pi = hosts.get(i).weight();
+            if (pi < 0) {
+                LOGGER.warn("{}: host at address {} has negative weight ({}). Using unweighted selection.",
+                        targetResource, hosts.get(i).address(), pi);
+                return new EqualWeightTable(hosts.size());
+            }
+            probs[i] = pi;
+            pTotal += pi;
+            allSameProbability = allSameProbability && approxEqual(pi, probs[0]);
+        }
+
+        if (allSameProbability) {
+            // no need for a DRV table.
+            return new EqualWeightTable(hosts.size());
+        }
+
+        // make sure our probability is normalized to less than 1% error.
+        if (!approxEqual(pTotal, 1)) {
+            // need to normalize.
+            double invPTotal = 1.0 / pTotal;
+            for (int i = 0; i < probs.length; i++) {
+                probs[i] *= invPTotal;
+            }
+        }
+        return buildAliasTable(probs);
+    }
+
+    private static abstract class EntrySelector {
+        abstract int firstEntry(Random random);
+        abstract int secondEntry(Random random, int firstEntry);
+    }
+
+    // An EntrySelector that represents equally weighted entries. In this case no alias table is needed.
+    private static final class EqualWeightTable extends EntrySelector {
+        private final int size;
+
+        EqualWeightTable(final int size) {
+            this.size = size;
+        }
+
+        @Override
+        int firstEntry(Random random) {
+            return random.nextInt(size);
+        }
+
+        @Override
+        int secondEntry(Random random, int firstEntry) {
+            assert size >= 2;
+            int result = random.nextInt(size - 1);
+            if (result >= firstEntry) {
+                result++;
+            }
+            return result;
+        }
+    }
+
+    // An EntrySelector that represents equally un-evenly weighted entries.
+    private final class AliasTableEntrySelector extends EntrySelector {
+
+        private final double[] aliasProbabilities;
+        private final int[] aliases;
+
+        AliasTableEntrySelector(double[] aliasProbabilities, int[] aliases) {
+            this.aliasProbabilities = aliasProbabilities;
+            this.aliases = aliases;
+        }
+
+        @Override
+        int secondEntry(Random random, int firstPick) {
+            int result;
+            int iteration = 0;
+            do {
+                result = pick(random);
+            } while (result == firstPick && iteration++ < maxEffort);
+            if (firstPick == result) {
+                LOGGER.debug("{}: failed to pick two unique indices after {} selection attempts",
+                        targetResource, maxEffort);
+            }
+            return 0;
+        }
+
+        @Override
+        int firstEntry(Random random) {
+            return pick(random);
+        }
+
+        private int pick(Random random) {
+            int i = random.nextInt(aliases.length);
+            double pp = aliasProbabilities[i];
+            if (pp != 1.0 && random.nextDouble() > pp) {
+                // use the alias.
+                i = aliases[i];
+            }
+            return i;
+        }
+    }
+
+    private AliasTableEntrySelector buildAliasTable(double[] pin) {
+        assert isNormalized(pin);
+        // We use the Vose alias method to compute the alias table with linear complexity.
+        // M. D. Vose, "A linear algorithm for generating random numbers with a given distribution,"
+        // IEEE Transactions on Software Engineering, vol. 17, no. 9, pp. 972-975, Sept. 1991, doi: 10.1109/32.92917.
+        // https://ieeexplore.ieee.org/document/92917
+        final double[] pout = new double[pin.length];
+        final int[] aliases = new int[pin.length];
+
+        // Setup our two stacks. Queues would also work but we'll use simple arrays as stacks.
+        final int[] small = new int[pin.length];
+        final int[] large = new int[pin.length];
+        int s = 0;
+        int l = 0;
+
+        for (int i = 0; i < pin.length; i++) {
+            pin[i] *= pin.length;
+            if (pin[i] > 1) {
+                large[l++] = i;
+            } else {
+                small[s++] = i;
+            }
+        }
+
+        // drain the stacks and populate the alias table
+        while (s != 0 && l != 0) {
+            int j = small[--s];
+            int k = large[--l];
+            pout[j] = pin[j];
+            aliases[j] = k;
+            pin[k] = pin[k] + pin[j] - 1;
+            if (pin[k] > 1) {
+                large[l++] = k;
+            } else {
+                small[s++] = k;
+            }
+        }
+        while (s != 0) {
+            pout[small[--s]] = 1;
+        }
+        while (l != 0) {
+            pout[large[--l]] = 1;
+        }
+        return new AliasTableEntrySelector(pout, aliases);
+    }
+
+    private static boolean approxEqual(double a, double b) {
+        double diff = a - b;
+        if (diff < 0) {
+            diff = -diff;
+        }
+        return diff < ACCEPTABLE_PERCENT_ERROR;
+    }
+
+    private static boolean isNormalized(double[] probabilities) {
+        double ptotal = 0;
+        for (double p : probabilities) {
+            ptotal += p;
+        }
+        return approxEqual(ptotal, 1);
     }
 }
