+++ servicetalk-loadbalancer-experimental/src/test/java/io/servicetalk/loadbalancer/P2CSelectorTest.java
@@ -16,19 +16,23 @@
 package io.servicetalk.loadbalancer;
 
 import io.servicetalk.client.api.NoActiveHostException;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
+import org.junit.jupiter.params.provider.CsvSource;
 
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Random;
 import java.util.concurrent.ExecutionException;
 import javax.annotation.Nullable;
 
 import static io.servicetalk.loadbalancer.SelectorTestHelpers.PREDICATE;
 import static io.servicetalk.loadbalancer.SelectorTestHelpers.connections;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.closeTo;
 import static org.hamcrest.Matchers.either;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.isA;
@@ -41,10 +45,12 @@ import static org.mockito.Mockito.when;
 
 class P2CSelectorTest {
 
+    private static final int ITERATIONS = 10_000;
+
     private boolean failOpen;
     private int maxEffort;
     @Nullable
-    private HostSelector<String, TestLoadBalancedConnection> selector;
+    private P2CSelector<String, TestLoadBalancedConnection> selector;
 
     @BeforeEach
     void setup() {
@@ -55,7 +61,52 @@ class P2CSelectorTest {
     }
 
     void init(List<Host<String, TestLoadBalancedConnection>> hosts) {
-        selector = new P2CSelector<>(hosts, "testResource", maxEffort, failOpen, null);
+        // Seed the random so that we don't have flaky test when we randomly get a bad roll.
+        selector = new P2CSelector<>(hosts, "testResource",
+                true, maxEffort, failOpen, new Random(1L));
+    }
+
+    @Test
+    void equalWeightDistribution() throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(5);
+        init(hosts);
+        checkProbabilities(hosts);
+    }
+
+    @Test
+    void unequalWeightDistribution() throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(5);
+        when(hosts.get(0).weight()).thenReturn(2.0);
+        when(hosts.get(1).weight()).thenReturn(0.5);
+        init(hosts);
+        checkProbabilities(hosts);
+    }
+
+    @Test
+    void unequalWeightsButWeightsDisabled() throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(2);
+        when(hosts.get(0).weight()).thenReturn(2.0);
+        when(hosts.get(1).weight()).thenReturn(0.5);
+        selector = new P2CSelector<>(hosts, "testResource", false, maxEffort, failOpen, new Random(0L));
+        int[] counts = runIterations(hosts);
+
+        double expected = Arrays.stream(counts).reduce(0, (a, b) -> a + b) / hosts.size();
+        for (int i = 0; i < hosts.size(); i++) {
+            assertThat((double) counts[i], closeTo(expected, expected * 0.05));
+        }
+    }
+
+    @Test
+    void negativeWeightsTurnIntoUnweightedSelection() throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(2);
+        when(hosts.get(0).weight()).thenReturn(-1.0);
+        init(hosts);
+        int[] counts = runIterations(hosts);
+
+        double expected = Arrays.stream(counts).reduce(0, (a, b) -> a + b) / hosts.size();
+        for (int i = 0; i < hosts.size(); i++) {
+            assertThat((double) counts[i], closeTo(expected, expected * 0.05));
+        }
     }
 
     @Test
@@ -66,10 +117,10 @@ class P2CSelectorTest {
         assertThat(connection.address(), equalTo("addr-1"));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void singleUnhealthyHost(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void singleUnhealthyHost(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1");
         when(hosts.get(0).isHealthy()).thenReturn(false);
         this.failOpen = failOpen;
         init(hosts);
@@ -84,10 +135,10 @@ class P2CSelectorTest {
         }
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void singleInactiveAndUnhealthyHostWithConnection(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void singleInactiveAndUnhealthyHostWithConnection(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1");
         when(hosts.get(0).isHealthy()).thenReturn(false);
         when(hosts.get(0).canMakeNewConnections()).thenReturn(false);
         this.failOpen = failOpen;
@@ -103,10 +154,10 @@ class P2CSelectorTest {
         }
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void singleInactiveAndUnhealthyHostWithoutConnection(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void singleInactiveAndUnhealthyHostWithoutConnection(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1");
         when(hosts.get(0).isHealthy()).thenReturn(false);
         when(hosts.get(0).canMakeNewConnections()).thenReturn(false);
         when(hosts.get(0).pickConnection(PREDICATE, null)).thenReturn(null);
@@ -118,19 +169,21 @@ class P2CSelectorTest {
         assertThat(e.getCause(), isA(NoActiveHostException.class));
     }
 
-    @Test
-    void twoHealthyActiveHosts() throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void twoHealthyActiveHosts(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
+        this.failOpen = failOpen;
         init(hosts);
         TestLoadBalancedConnection connection = selector.selectConnection(
                 PREDICATE, null, true).toFuture().get();
         assertThat(connection.address(), either(equalTo("addr-1")).or(equalTo("addr-2")));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void twoHealthyInactiveHostsWithConnections(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void twoHealthyInactiveHostsWithConnections(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         for (Host<String, TestLoadBalancedConnection> host : hosts) {
             when(host.canMakeNewConnections()).thenReturn(false);
         }
@@ -141,10 +194,10 @@ class P2CSelectorTest {
         assertThat(connection.address(), either(equalTo("addr-1")).or(equalTo("addr-2")));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void twoHealthyInactiveHostsWithoutConnections(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void twoHealthyInactiveHostsWithoutConnections(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         for (Host<String, TestLoadBalancedConnection> host : hosts) {
             when(host.canMakeNewConnections()).thenReturn(false);
             when(host.pickConnection(PREDICATE, null)).thenReturn(null);
@@ -156,10 +209,10 @@ class P2CSelectorTest {
         assertThat(e.getCause(), isA(NoActiveHostException.class));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void twoUnHealthyActiveHosts(boolean failOpen) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void twoUnHealthyActiveHosts(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         for (Host<String, TestLoadBalancedConnection> host : hosts) {
             when(host.isHealthy()).thenReturn(false);
         }
@@ -176,10 +229,10 @@ class P2CSelectorTest {
         }
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}")
-    @ValueSource(booleans = {false, true})
-    void twoUnHealthyInactiveHosts(boolean failOpen) {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void twoUnHealthyInactiveHosts(boolean failOpen, boolean equalWeights) {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         for (Host<String, TestLoadBalancedConnection> host : hosts) {
             when(host.isHealthy()).thenReturn(false);
             when(host.canMakeNewConnections()).thenReturn(false);
@@ -191,12 +244,14 @@ class P2CSelectorTest {
         assertThat(e.getCause(), isA(NoActiveHostException.class));
     }
 
-    @Test
-    void doesntBiasTowardHostsWithConnections() throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: failOpen={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void doesntBiasTowardHostsWithConnections(boolean failOpen, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         // we setup the first host to always be preferred by score, but it also doesn't have any connections.
         when(hosts.get(0).pickConnection(any(), any())).thenReturn(null);
         when(hosts.get(0).score()).thenReturn(10);
+        this.failOpen = failOpen;
         init(hosts);
         TestLoadBalancedConnection connection = selector.selectConnection(
                 PREDICATE, null, false).toFuture().get();
@@ -231,10 +286,10 @@ class P2CSelectorTest {
         assertThat(connection.address(), equalTo("addr-2"));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: forceNewConnection={0}")
-    @ValueSource(booleans = {false, true})
-    void biasesTowardTheHighestWeightHost(boolean forceNewConnection) throws Exception {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1", "addr-2");
+    @ParameterizedTest(name = "{displayName} [{index}]: forceNewConnection={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void biasesTowardTheHighestWeightHost(boolean forceNewConnection, boolean equalWeights) throws Exception {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1", "addr-2");
         // Host 0 has the highest score, so it should always get the new connection.
         when(hosts.get(0).score()).thenReturn(10);
         init(hosts);
@@ -243,10 +298,10 @@ class P2CSelectorTest {
         assertThat(connection.address(), equalTo("addr-1"));
     }
 
-    @ParameterizedTest(name = "{displayName} [{index}]: unhealthy={0}")
-    @ValueSource(booleans = {false, true})
-    void singleInactiveHostFailOpen(boolean unhealthy) {
-        List<Host<String, TestLoadBalancedConnection>> hosts = connections("addr-1");
+    @ParameterizedTest(name = "{displayName} [{index}]: unhealthy={0}, equalWeights={1}")
+    @CsvSource({"true, true,", "true, false", "false, true", "false, false"})
+    void singleInactiveHostFailOpen(boolean unhealthy, boolean equalWeights) {
+        List<Host<String, TestLoadBalancedConnection>> hosts = connections(equalWeights, "addr-1");
         when(hosts.get(0).isHealthy()).thenReturn(unhealthy);
         when(hosts.get(0).canMakeNewConnections()).thenReturn(false);
         when(hosts.get(0).pickConnection(PREDICATE, null)).thenReturn(null);
@@ -255,5 +310,40 @@ class P2CSelectorTest {
         Exception e = assertThrows(ExecutionException.class, () -> selector.selectConnection(
                 PREDICATE, null, false).toFuture().get());
         assertThat(e.getCause(), isA(NoActiveHostException.class));
+    }
+
+    private void checkProbabilities(List<Host<String, TestLoadBalancedConnection>> hosts) throws Exception {
+        int[] counts = runIterations(hosts);
+
+        double totalProbability = hosts.stream().map(Host::weight).reduce(0d, (a, b) -> a + b);
+        Integer[] expected = hosts.stream().map(host -> (int)(ITERATIONS * (host.weight() / totalProbability)))
+                .toArray(Integer[]::new);
+
+        // calculate the rough counts we should expect
+        for (int i = 0; i < hosts.size(); i++) {
+            double c = counts[i];
+            double e = expected[i];
+            double acceptableError = 0.05 * e; // 5% error
+            assertThat(c, closeTo(e, acceptableError));
+        }
+    }
+
+    private int[] runIterations(List<Host<String, TestLoadBalancedConnection>> hosts) throws Exception {
+        HostSelector<String, TestLoadBalancedConnection> toRun = selector.rebuildWithHosts(hosts);
+        int[] results = new int[hosts.size()];
+        Map<TestLoadBalancedConnection, Integer> hostMap = new HashMap<>();
+
+        // setup our hosts with a connection
+        for (int i = 0; i < hosts.size(); i++) {
+            Host<String, TestLoadBalancedConnection> host = hosts.get(i);
+            TestLoadBalancedConnection cxn = TestLoadBalancedConnection.mockConnection(host.address());
+            when(host.pickConnection(any(), any())).thenReturn(cxn);
+            hostMap.put(cxn, i);
+        }
+
+        for (int i = 0; i < ITERATIONS; i++) {
+            results[hostMap.get(toRun.selectConnection(ignored -> true, null, false).toFuture().get())]++;
+        }
+        return results;
     }
 }
